+++ drivers/lcd.py
@@ -50,7 +50,7 @@ En = 0b00000100 # Enable bit
 Rw = 0b00000010 # Read/Write bit
 Rs = 0b00000001 # Register select bit
 
-class lcd:
+class LCD:
    #initializes objects and lcd
    line = 1
    def __init__(self):
